+++ pkg/storage/tsdb/config.go
@@ -104,17 +104,19 @@ func (cfg *Config) Validate() error {
 
 // BucketStoreConfig holds the config information for Bucket Stores used by the querier
 type BucketStoreConfig struct {
-	SyncDir              string `yaml:"sync_dir"`
-	IndexCacheSizeBytes  uint64 `yaml:"index_cache_size_bytes"`
-	MaxChunkPoolBytes    uint64 `yaml:"max_chunk_pool_bytes"`
-	MaxSampleCount       uint64 `yaml:"max_sample_count"`
-	MaxConcurrent        int    `yaml:"max_concurrent"`
-	BlockSyncConcurrency int    `yaml:"block_sync_concurrency"`
+	SyncDir              string        `yaml:"sync_dir"`
+	SyncInterval         time.Duration `yaml:"sync_interval"`
+	IndexCacheSizeBytes  uint64        `yaml:"index_cache_size_bytes"`
+	MaxChunkPoolBytes    uint64        `yaml:"max_chunk_pool_bytes"`
+	MaxSampleCount       uint64        `yaml:"max_sample_count"`
+	MaxConcurrent        int           `yaml:"max_concurrent"`
+	BlockSyncConcurrency int           `yaml:"block_sync_concurrency"`
 }
 
 // RegisterFlags registers the BucketStore flags
 func (cfg *BucketStoreConfig) RegisterFlags(f *flag.FlagSet) {
 	f.StringVar(&cfg.SyncDir, "experimental.tsdb.bucket-store.sync-dir", "tsdb-sync", "Directory to place synced tsdb indicies.")
+	f.DurationVar(&cfg.SyncInterval, "experimental.tsdb.bucket-store.sync-interval", 5*time.Minute, "How frequently scan the bucket to look for changes (new blocks shipped by ingesters and blocks removed by retention or compaction). 0 disables it.")
 	f.Uint64Var(&cfg.IndexCacheSizeBytes, "experimental.tsdb.bucket-store.index-cache-size-bytes", uint64(250*units.Mebibyte), "Size of index cache in bytes per tenant.")
 	f.Uint64Var(&cfg.MaxChunkPoolBytes, "experimental.tsdb.bucket-store.max-chunk-pool-bytes", uint64(2*units.Gibibyte), "Max size of chunk pool in bytes per tenant.")
 	f.Uint64Var(&cfg.MaxSampleCount, "experimental.tsdb.bucket-store.max-sample-count", 0, "Max number of samples (0 is no limit) per query when loading series from storage.")